Move game-parser to its own package
Simple move class refactor and change method visibility to
support the class move.
