+++ game-core/src/main/java/games/strategy/engine/data/ProductionRule.java
@@ -25,12 +25,12 @@ public class ProductionRule extends DefaultNamed {
     this.costs = costs;
   }
 
-  protected void addCost(final Resource resource, final int quantity) {
+  public void addCost(final Resource resource, final int quantity) {
     costs.put(resource, quantity);
   }
 
   /** Benefits must be a resource or a unit. */
-  protected void addResult(final NamedAttachable obj, final int quantity) {
+  public void addResult(final NamedAttachable obj, final int quantity) {
     if (!(obj instanceof UnitType) && !(obj instanceof Resource)) {
       throw new IllegalArgumentException(
           "results must be units or resources, not:" + obj.getClass().getName());
