+++ game-core/src/main/java/games/strategy/engine/framework/ui/DefaultGameChooserEntry.java
@@ -3,7 +3,7 @@ package games.strategy.engine.framework.ui;
 import games.strategy.engine.data.EngineVersionException;
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.GameParseException;
-import games.strategy.engine.data.GameParser;
+import games.strategy.engine.data.gameparser.GameParser;
 import games.strategy.triplea.Constants;
 import java.io.IOException;
 import java.io.InputStream;
